+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+* Show DelphiLint actions in IDE Insight search bar
+* Allow DelphiLint actions to be added to editor toolbars
+
 ## [0.1.0] - 2023-07-20
 
 ### Added