+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 2023-07-25
+
 ### Added
 
 * Show DelphiLint actions in IDE Insight search bar
@@ -33,4 +35,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Support for reading `sonar-project.properties`
 
 [unreleased]: https://github.com/Integrated-Application-Development/delphilint/compare/v0.1.0...HEAD
+[0.2.0]: https://github.com/Integrated-Application-Development/delphilint/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/Integrated-Application-Development/delphilint/releases/tag/v0.1.0